requirements.txt
nbconvert
nbformat
numpy
scikit-learn
opencv-python
Pillow
scipy
pandas
streamlit